Refactor net module to one include directory
