+++ drivers/cc110x/Makefile
@@ -1,6 +1,6 @@
 MODULE =cc110x
 
-INCLUDES += -I$(RIOTBASE)/sys/net
+INCLUDES += -I$(RIOTBASE)/sys/net/include
 
 include $(MAKEBASE)/Makefile.base
