+++ drivers/cc110x/cc1100-csmaca-mac.c
@@ -32,7 +32,7 @@ and Telematics group (http://cst.mi.fu-berlin.de).
 #include <cc1100.h>
 #include <cc1100_phy.h>
 #include <cc1100-csmaca-mac.h>
-#include "protocol-multiplex/protocol-multiplex.h"
+#include "protocol-multiplex.h"
 
 #include "hwtimer.h"
 #include <vtimer.h>
