+++ sys/net/destiny/Makefile
@@ -1,4 +1,4 @@
 MODULE:=$(shell basename $(CURDIR))
-INCLUDES = -I$(RIOTBASE)/sys/include -I$(RIOTBASE)/core/include -I$(RIOTBASE)/drivers/include -I$(RIOTBASE)/sys/net/destiny/include -I$(RIOTBASE)/sys/net/sixlowpan/include -I$(RIOTBASE)/sys/net/net_help/ -I$(RIOTBASE)/drivers/cc110x_ng/include  -I$(RIOTBASE)/drivers/cc110x
+INCLUDES = -I$(RIOTBASE)/sys/include -I$(RIOTBASE)/core/include -I$(RIOTBASE)/drivers/include -I$(RIOTBASE)/sys/net/include -I$(RIOTBASE)/drivers/cc110x_ng/include  -I$(RIOTBASE)/drivers/cc110x
 
 include $(RIOTBASE)/Makefile.base
