+++ sys/net/destiny/socket.c
@@ -26,7 +26,7 @@
 #include "thread.h"
 #include "vtimer.h"
 
-#include "../net_help/net_help.h"
+#include "net_help.h"
 
 #include "msg_help.h"
 #include "tcp.h"
