+++ sys/net/destiny/tcp.c
@@ -26,7 +26,7 @@
 
 #include "destiny/in.h"
 
-#include "../net_help/net_help.h"
+#include "net_help.h"
 
 #include "msg_help.h"
 #include "socket.h"
