+++ sys/net/ieee802154/Makefile
@@ -1,4 +1,4 @@
 MODULE:=$(shell basename $(CURDIR))
-INCLUDES = -I$(RIOTBASE) -I$(RIOTBASE)/sys/include -I$(RIOTBASE)/core/include -I$(RIOTBASE)/drivers/include -I$(RIOTBASE)/sys/net -I$(RIOTBASE)/sys/net/ieee802154/include
+INCLUDES = -I$(RIOTBASE) -I$(RIOTBASE)/sys/include -I$(RIOTBASE)/core/include -I$(RIOTBASE)/drivers/include -I$(RIOTBASE)/sys/net/include
 
 include $(RIOTBASE)/Makefile.base
