+++ sys/net/net_help/Makefile
@@ -1,4 +1,4 @@
 MODULE:=$(shell basename $(CURDIR))
-INCLUDES += -I$(RIOTBASE)/sys/net/destiny/include -I$(RIOTBASE)/sys/include -I$(RIOTBASE)/core/include -I$(RIOTBASE)/sys/net/sixlowpan/include
+INCLUDES += -I$(RIOTBASE)/sys/net/include -I$(RIOTBASE)/sys/include -I$(RIOTBASE)/core/include
 
 include $(RIOTBASE)/Makefile.base
