+++ sys/net/protocol-multiplex/Makefile
@@ -1,4 +1,4 @@
 MODULE:=protocol_multiplex
-INCLUDES = -I$(RIOTBASE) -I$(RIOTBASE)/sys/include -I$(RIOTBASE)/core/include
+INCLUDES = -I$(RIOTBASE) -I$(RIOTBASE)/sys/include -I$(RIOTBASE)/core/include -I$(RIOTBASE)/sys/net/include
 
 include $(RIOTBASE)/Makefile.base
